python/python_flask/requirements.txt
flask
pillow
numpy
sentence-transformers
google-generativeai
faiss-cpu
pymysql
werkzeug
